Move astrowidgets API to mixin class and use it for viewer, not Imviz instance.
Change marker size in example notebook to reflect UI max.

Disallow viewer from removing markers added by another viewer.

Clarify how markers play with relinking.
