+++ docs/imviz/notebook.rst
@@ -12,7 +12,8 @@ To run Imviz in a notebook::
 
 Imviz also provides programmatic access to its viewers using
 `Astrowidgets <https://astrowidgets.readthedocs.io/en/latest/>`_ API.
-For example::
+See `~jdaviz.core.astrowidgets_api.AstrowidgetsImageViewerMixin` for
+available functionality. For example::
 
     viewer = imviz.app.get_viewer('viewer-1')
     viewer.center_on((100, 100))
