+++ jdaviz/configs/imviz/tests/test_astrowidgets_api.py
@@ -16,7 +16,7 @@ class TestSave(BaseImviz_WCS_NoWCS):
 
     def test_save(self, tmpdir):
         filename = os.path.join(tmpdir.strpath, 'myimage')
-        self.imviz.save(filename)
+        self.viewer.save(filename)
 
         # This only tests that something saved, not the content.
         assert os.path.isfile(os.path.join(tmpdir.strpath, 'myimage.png'))
@@ -25,33 +25,33 @@ class TestSave(BaseImviz_WCS_NoWCS):
 class TestCenterOffset(BaseImviz_WCS_NoWCS):
 
     def test_center_offset_pixel(self):
-        self.imviz.center_on((0, 1))
+        self.viewer.center_on((0, 1))
         assert_allclose(self.viewer.state.x_min, -5)
         assert_allclose(self.viewer.state.y_min, -4)
         assert_allclose(self.viewer.state.x_max, 5)
         assert_allclose(self.viewer.state.y_max, 6)
 
-        self.imviz.offset_by(1 * u.pix, -1 * u.dimensionless_unscaled)
+        self.viewer.offset_by(1 * u.pix, -1 * u.dimensionless_unscaled)
         assert_allclose(self.viewer.state.x_min, -4)
         assert_allclose(self.viewer.state.y_min, -5)
         assert_allclose(self.viewer.state.x_max, 6)
         assert_allclose(self.viewer.state.y_max, 5)
 
-        self.imviz.offset_by(1, 0)
+        self.viewer.offset_by(1, 0)
         assert_allclose(self.viewer.state.x_min, -3)
         assert_allclose(self.viewer.state.y_min, -5)
         assert_allclose(self.viewer.state.x_max, 7)
         assert_allclose(self.viewer.state.y_max, 5)
 
         # Out-of-bounds centering should be no-op
-        self.imviz.center_on((-1, 99999))
+        self.viewer.center_on((-1, 99999))
         assert_allclose(self.viewer.state.x_min, -3)
         assert_allclose(self.viewer.state.y_min, -5)
         assert_allclose(self.viewer.state.x_max, 7)
         assert_allclose(self.viewer.state.y_max, 5)
 
         # Sometimes invalid WCS also gives such output, should be no-op
-        self.imviz.center_on((np.array(np.nan), np.array(np.nan)))
+        self.viewer.center_on((np.array(np.nan), np.array(np.nan)))
         assert_allclose(self.viewer.state.x_min, -3)
         assert_allclose(self.viewer.state.y_min, -5)
         assert_allclose(self.viewer.state.x_max, 7)
@@ -62,14 +62,14 @@ class TestCenterOffset(BaseImviz_WCS_NoWCS):
         self.viewer.blink_once()
 
         sky = self.wcs.pixel_to_world(0, 1)
-        self.imviz.center_on(sky)
+        self.viewer.center_on(sky)
         assert_allclose(self.viewer.state.x_min, -5)
         assert_allclose(self.viewer.state.y_min, -4)
         assert_allclose(self.viewer.state.x_max, 5)
         assert_allclose(self.viewer.state.y_max, 6)
 
         dsky = 0.1 * u.arcsec
-        self.imviz.offset_by(-dsky, dsky)
+        self.viewer.offset_by(-dsky, dsky)
         assert_allclose(self.viewer.state.x_min, -4.9)
         assert_allclose(self.viewer.state.y_min, -3.90000000002971)
         assert_allclose(self.viewer.state.x_max, 5.1)
@@ -77,22 +77,22 @@ class TestCenterOffset(BaseImviz_WCS_NoWCS):
 
         # Cannot mix pixel with sky
         with pytest.raises(ValueError, match='but dy is of type'):
-            self.imviz.offset_by(0.1, dsky)
+            self.viewer.offset_by(0.1, dsky)
 
         # Cannot pass invalid Quantity
         with pytest.raises(u.UnitTypeError):
-            self.imviz.offset_by(dsky, 1 * u.AA)
+            self.viewer.offset_by(dsky, 1 * u.AA)
         with pytest.raises(u.UnitTypeError):
-            self.imviz.offset_by(1 * u.AA, dsky)
+            self.viewer.offset_by(1 * u.AA, dsky)
 
         # Blink to the one without WCS
         self.viewer.blink_once()
 
         with pytest.raises(AttributeError, match='does not have a valid WCS'):
-            self.imviz.center_on(sky)
+            self.viewer.center_on(sky)
 
         with pytest.raises(AttributeError, match='does not have a valid WCS'):
-            self.imviz.offset_by(dsky, dsky)
+            self.viewer.offset_by(dsky, dsky)
 
 
 class TestZoom(BaseImviz_WCS_NoWCS):
@@ -100,14 +100,14 @@ class TestZoom(BaseImviz_WCS_NoWCS):
     @pytest.mark.parametrize('val', (0, -0.1, 'foo', [1, 2]))
     def test_invalid_zoom_level(self, val):
         with pytest.raises(ValueError, match='Unsupported zoom level'):
-            self.imviz.zoom_level = val
+            self.viewer.zoom_level = val
 
     def test_invalid_zoom(self):
         with pytest.raises(ValueError, match='zoom only accepts int or float'):
-            self.imviz.zoom('fit')
+            self.viewer.zoom('fit')
 
     def assert_zoom_results(self, zoom_level, x_min, x_max, y_min, y_max, dpix):
-        assert_allclose(self.imviz.zoom_level, zoom_level)
+        assert_allclose(self.viewer.zoom_level, zoom_level)
         assert_allclose((self.viewer.state.x_min, self.viewer.state.x_max,
                          self.viewer.state.y_min, self.viewer.state.y_max),
                         (x_min + dpix, x_max + dpix,
@@ -116,51 +116,51 @@ class TestZoom(BaseImviz_WCS_NoWCS):
     @pytest.mark.parametrize('is_offcenter', (False, True))
     def test_zoom(self, is_offcenter):
         if is_offcenter:
-            self.imviz.center_on((0, 0))
+            self.viewer.center_on((0, 0))
             dpix = -4.5
         else:
-            self.imviz.center_on((4.5, 4.5))
+            self.viewer.center_on((4.5, 4.5))
             dpix = 0
 
         self.assert_zoom_results(10, -0.5, 9.5, -0.5, 9.5, dpix)
 
         # NOTE: Not sure why X/Y min/max not exactly the same as aspect ratio 1
-        self.imviz.zoom_level = 1
+        self.viewer.zoom_level = 1
         self.assert_zoom_results(1, -46, 54, -45.5, 54.5, dpix)
 
-        self.imviz.zoom_level = 2
+        self.viewer.zoom_level = 2
         self.assert_zoom_results(2, -21.5, 28.5, -20.5, 29.5, dpix)
 
-        self.imviz.zoom(2)
+        self.viewer.zoom(2)
         self.assert_zoom_results(4, -9.5, 15.5, -8.0, 17.0, dpix)
 
-        self.imviz.zoom(0.5)
+        self.viewer.zoom(0.5)
         self.assert_zoom_results(2, -22.5, 27.5, -20.5, 29.5, dpix)
 
-        self.imviz.zoom_level = 0.5
+        self.viewer.zoom_level = 0.5
         self.assert_zoom_results(0.5, -98, 102, -95.5, 104.5, dpix)
 
         # This fits the whole image on screen, regardless.
         # NOTE: But somehow Y min/max auto-adjust does not work properly
         # in the unit test when off-center. Works in notebook though.
         if not is_offcenter:
-            self.imviz.zoom_level = 'fit'
+            self.viewer.zoom_level = 'fit'
             self.assert_zoom_results(10, -0.5, 9.5, -0.5, 9.5, 0)
 
 
 class TestCmapStretchCuts(BaseImviz_WCS_NoWCS):
 
     def test_colormap_options(self):
-        assert self.imviz.colormap_options == [
+        assert self.viewer.colormap_options == [
             'BuGn', 'PRGn', 'PuBu', 'PuOr', 'RdBu', 'RdPu', 'RdYlBu', 'YlGnBu', 'YlOrRd',
             'gray', 'hot', 'inferno', 'magma', 'plasma', 'viridis']
 
     def test_invalid_colormap(self):
         with pytest.raises(ValueError, match='Invalid colormap'):
-            self.imviz.set_colormap('foo')
+            self.viewer.set_colormap('foo')
 
     def test_stretch_options(self):
-        assert self.imviz.stretch_options == ['arcsinh', 'linear', 'log', 'sqrt']
+        assert self.viewer.stretch_options == ['arcsinh', 'linear', 'log', 'sqrt']
 
     @pytest.mark.parametrize(('vizclass', 'ans'),
                              [(AsinhStretch, 'arcsinh'),
@@ -168,21 +168,21 @@ class TestCmapStretchCuts(BaseImviz_WCS_NoWCS):
                               (LogStretch, 'log'),
                               (SqrtStretch, 'sqrt')])
     def test_stretch_astropy(self, vizclass, ans):
-        self.imviz.stretch = vizclass
-        assert self.imviz.stretch == ans
+        self.viewer.stretch = vizclass
+        assert self.viewer.stretch == ans
 
     def test_invalid_stretch(self):
         class FakeStretch:
             pass
 
         with pytest.raises(ValueError, match='Invalid stretch'):
-            self.imviz.stretch = FakeStretch
+            self.viewer.stretch = FakeStretch
 
         with pytest.raises(ValueError, match='Invalid stretch'):
-            self.imviz.stretch = 'foo'
+            self.viewer.stretch = 'foo'
 
     def test_autocut_options(self):
-        assert self.imviz.autocut_options == ['minmax', '99.5%', '99%', '95%', '90%']
+        assert self.viewer.autocut_options == ['minmax', '99.5%', '99%', '95%', '90%']
 
     @pytest.mark.parametrize(('auto_option', 'ans'),
                              [('minmax', (0, 99)),
@@ -191,30 +191,30 @@ class TestCmapStretchCuts(BaseImviz_WCS_NoWCS):
                               ('95%', (2.475, 96.525)),
                               ('90%', (4.95, 94.05))])
     def test_autocut(self, auto_option, ans):
-        self.imviz.cuts = auto_option
-        assert_allclose(self.imviz.cuts, ans)
+        self.viewer.cuts = auto_option
+        assert_allclose(self.viewer.cuts, ans)
 
     def test_invalid_autocut(self):
         with pytest.raises(ValueError, match='Invalid autocut'):
-            self.imviz.cuts = 'foo'
+            self.viewer.cuts = 'foo'
 
     @pytest.mark.parametrize('val', [99, (1, ), (1, 2, 3), (1, 'foo')])
     def test_invalid_cuts(self, val):
         with pytest.raises(ValueError, match='Invalid cut levels'):
-            self.imviz.cuts = val
+            self.viewer.cuts = val
 
     def test_cmap_stretch_cuts(self):
         # Change colormap, stretch, and cuts on one image
-        self.imviz.set_colormap('viridis')
-        self.imviz.stretch = 'sqrt'
-        self.imviz.cuts = '95%'
+        self.viewer.set_colormap('viridis')
+        self.viewer.stretch = 'sqrt'
+        self.viewer.cuts = '95%'
 
         self.viewer.blink_once()
 
         # Change colormap, stretch, and cuts on other image
-        self.imviz.set_colormap('RdYlBu')
-        self.imviz.stretch = AsinhStretch
-        self.imviz.cuts = (0, 100)
+        self.viewer.set_colormap('RdYlBu')
+        self.viewer.stretch = AsinhStretch
+        self.viewer.cuts = (0, 100)
 
         # Make sure settings stick on both images, second image displayed/changed first above.
         assert self.viewer.state.layers[0].cmap.name == 'RdYlBu'
@@ -235,15 +235,15 @@ class TestMarkers(BaseImviz_WCS_NoWCS):
 
     def test_invalid_markers(self):
         with pytest.raises(KeyError, match='Invalid attribute'):
-            self.imviz.marker = {'foo': 'bar', 'alpha': 0.8}
+            self.viewer.marker = {'foo': 'bar', 'alpha': 0.8}
         with pytest.raises(ValueError, match='Invalid RGBA argument'):
-            self.imviz.marker = {'color': 'greenfishbluefish'}
+            self.viewer.marker = {'color': 'greenfishbluefish'}
         with pytest.raises(ValueError, match='Invalid alpha'):
-            self.imviz.marker = {'alpha': '1'}
+            self.viewer.marker = {'alpha': '1'}
         with pytest.raises(ValueError, match='Invalid alpha'):
-            self.imviz.marker = {'alpha': 42}
+            self.viewer.marker = {'alpha': 42}
         with pytest.raises(ValueError, match='Invalid marker size'):
-            self.imviz.marker = {'markersize': '1'}
+            self.viewer.marker = {'markersize': '1'}
 
     def test_mvp_markers(self):
         x_pix = (0, 0)
@@ -251,7 +251,7 @@ class TestMarkers(BaseImviz_WCS_NoWCS):
         sky = self.wcs.pixel_to_world(x_pix, y_pix)
         tbl = Table({'x': x_pix, 'y': y_pix, 'coord': sky})
 
-        self.imviz.add_markers(tbl)
+        self.viewer.add_markers(tbl)
         data = self.imviz.app.data_collection[2]
         assert data.label == 'default-marker-name'
         assert data.style.color == 'red'
@@ -263,15 +263,15 @@ class TestMarkers(BaseImviz_WCS_NoWCS):
 
         # Table with only sky coordinates but no use_skycoord=True
         with pytest.raises(KeyError):
-            self.imviz.add_markers(tbl[('coord', )])
+            self.viewer.add_markers(tbl[('coord', )])
 
         # Cannot use reserved marker name
         with pytest.raises(ValueError, match='not allowed'):
-            self.imviz.add_markers(tbl, use_skycoord=True, marker_name='all')
+            self.viewer.add_markers(tbl, use_skycoord=True, marker_name='all')
 
-        self.imviz.marker = {'color': (0, 1, 0), 'alpha': 0.8}
+        self.viewer.marker = {'color': (0, 1, 0), 'alpha': 0.8}
 
-        self.imviz.add_markers(tbl, use_skycoord=True, marker_name='my_sky')
+        self.viewer.add_markers(tbl, use_skycoord=True, marker_name='my_sky')
         data = self.imviz.app.data_collection[3]
         assert data.label == 'my_sky'
         assert data.style.color == (0, 1, 0)  # green
@@ -285,12 +285,12 @@ class TestMarkers(BaseImviz_WCS_NoWCS):
         assert len(self.imviz.app.data_collection.links) == 14
 
         # Remove markers with default name.
-        self.imviz.remove_markers()
+        self.viewer.remove_markers()
         assert self.imviz.app.data_collection.labels == [
             'has_wcs[SCI,1]', 'no_wcs[SCI,1]', 'my_sky']
 
         # Reset markers (runs remove_markers with marker_name set)
-        self.imviz.reset_markers()
+        self.viewer.reset_markers()
         assert self.imviz.app.data_collection.labels == [
             'has_wcs[SCI,1]', 'no_wcs[SCI,1]']
 
@@ -300,8 +300,8 @@ class TestMarkers(BaseImviz_WCS_NoWCS):
 
         self.imviz.app.data_collection.remove(self.imviz.app.data_collection[0])
         with pytest.raises(AttributeError, match='does not have a valid WCS'):
-            self.imviz.add_markers(tbl, use_skycoord=True, marker_name='my_sky')
+            self.viewer.add_markers(tbl, use_skycoord=True, marker_name='my_sky')
 
         self.imviz.app.data_collection.clear()
         with pytest.raises(AttributeError, match='does not have a valid WCS'):
-            self.imviz.add_markers(tbl, use_skycoord=True, marker_name='my_sky')
+            self.viewer.add_markers(tbl, use_skycoord=True, marker_name='my_sky')
